utils/aisearch.py
import os
from azure.core.credentials import AzureKeyCredential
from azure.search.documents import SearchClient

service_endpoint = os.environ["AZURE_SEARCH_SERVICE_ENDPOINT"]
index_name = os.environ["AZURE_SEARCH_INDEX_NAME"]
key = os.environ["AZURE_SEARCH_API_KEY"]

search_client = SearchClient(service_endpoint, index_name, AzureKeyCredential(key))


def retrieval(search_text):
    results = search_client.search(search_text=search_text)
    return results[0]["content"]


if __name__ == "__main__":
    print("Searching")