requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
# aenum==3.1.15
# aiohttp==3.9.3
# aiosignal==1.3.1
# annotated-types==0.6.0
# attrs==23.2.0
# azure-common==1.1.28
# azure-core==1.30.0
# azure-cosmos==4.5.1
# azure-identity==1.15.0
# azure-search-documents==11.4.0
# beautifulsoup4==4.12.3
# certifi==2024.2.2
# cffi==1.16.0
# charset-normalizer==3.3.2
# cryptography==42.0.2
# Deprecated==1.2.14
# frozenlist==1.4.1
# future==0.18.3
# idna==3.6
# isodate==0.6.1
line-bot-sdk==3.8.1
# lxml==5.1.0
# msal==1.26.0
# msal-extensions==1.1.0
# multidict==6.0.5
# packaging==23.2
# portalocker==2.8.2
# pycparser==2.21
# pydantic==2.6.1
# pydantic_core==2.16.2
# PyJWT==2.8.0
# python-dateutil==2.8.2
# pywin32==306
# requests==2.31.0
# six==1.16.0
# soupsieve==2.5
# typing_extensions==4.9.0
# urllib3==2.2.0
# wrapt==1.16.0
# yarl==1.9.4